Remove old protocol files, add new ones
Deleted the old protocol files to remove dependency on the old implementation. New files were added with updated services and schemas to align with the new implementation in the Houston Gradle plugin. Renamed and moved Manifest's proto file to adhere to the updated folder structure. Also, included a new mapping in the VCS for version control purposes.