Updated gradle dependencies, Reinforced Error Handling and Added Data Synchronization Task
A variety of changes were made, primarily centered around error handling, data synchronization, and updated dependencies. The "Plugin.java" file was deleted due to it being a placeholder, and the HoustonEnvironment.java file was added to handle environment settings for a Houston instance.

In AssetContainerWatcher.java, an error resume has been added so that if downloadArchive fails, an error log will be logged and the operation will continue. ColumnType method in AssetContainer.java now does not need a Scope parameter. UploadCommand.java was changed to simplify and fix issues.

Updated versions of protobuf, grpc, and lombok are reflected in gradle.properties.

Dependencies and plugins have been adjusted and added in build.gradle.

Finally, the SyncSchema task was added in SyncSchema.java. This task synchronizes schema by fetching it from a given URL, and writes it to the local system. The fetched schema is written in the standard .proto format, making it compatible with other systems and languages that support protocol buffers.

These changes should improve the reliability of our builds and synchronization. It also introduce new features like data synchronization and environment settings.