Refactor code and update RPC method name
The codebase has been refactored which includes some reordering of methods, style corrections, and deletion of the class 'PublishVerifier'. In addition, the name of the RPC method in PluginService.proto has been modified from 'runVerifierRequest' to 'UploadVerifierRequest' for better clarity. Also, the 'HoustonEnvironment' class now directly holds a 'verifier' property of type Project. This commit also includes the correction of some protos to improve the code's efficiency and readability.